src/requirements.txt
numpy~=1.19.5
opencv-python~=4.5.1.48
matplotlib~=3.3.4
Pillow~=8.1.0